Rename folder to `ExecutionPathTracer`
