+++ src/ExecutionPathTracer/ExecutionPath.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Microsoft.Quantum.Simulation.Core;
 
-namespace Microsoft.Quantum.IQSharp.Tracer
+namespace Microsoft.Quantum.IQSharp.ExecutionPathTracer
 {
     /// <summary>
     /// Represents the qubit resources and operations traced out in an execution path of a Q# operation.
