+++ src/ExecutionPathTracer/ExecutionPathTracer.csproj
@@ -3,8 +3,8 @@
   <PropertyGroup>
     <TargetFramework>netstandard2.1</TargetFramework>
     <PlatformTarget>x64</PlatformTarget>
-    <RootNamespace>Microsoft.Quantum.IQSharp.Tracer</RootNamespace>
-    <AssemblyName>Microsoft.Quantum.IQSharp.Tracer</AssemblyName>
+    <RootNamespace>Microsoft.Quantum.IQSharp.ExecutionPathTracer</RootNamespace>
+    <AssemblyName>Microsoft.Quantum.IQSharp.ExecutionPathTracer</AssemblyName>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
   </PropertyGroup>
