+++ src/ExecutionPathTracer/Register.cs
@@ -6,7 +6,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
-namespace Microsoft.Quantum.IQSharp.Tracer
+namespace Microsoft.Quantum.IQSharp.ExecutionPathTracer
 {
     /// <summary>
     /// Enum for the 2 types of registers: <c>Qubit</c> and <c>Classical</c>.
