+++ src/Kernel/Kernel.csproj
@@ -40,7 +40,7 @@
   <ItemGroup>
     <ProjectReference Include="..\AzureClient\AzureClient.csproj" />
     <ProjectReference Include="..\Core\Core.csproj" />
-    <ProjectReference Include="..\Tracer\Tracer.csproj" />
+    <ProjectReference Include="..\ExecutionPathTracer\ExecutionPathTracer.csproj" />
     <ProjectReference Include="..\Jupyter\Jupyter.csproj" />
   </ItemGroup>
